mainLogic正式下岗
