--- GTMBplugin_B/Script_NeteaseMod_TtCLQKGx/Parts/MainLogic/MainLogic.part
@@ -1,11 +1,0 @@
-{
-    "bpType": "MainLogicPart",
-    "changes": {},
-    "classType": "PartInstance",
-    "clientPartType": null,
-    "customVariables": [],
-    "directoryPath": "MainLogic",
-    "metaType": "MainLogicPartMeta",
-    "objId": "ce01f37f3e394aa2a2d7a871a0ddb226",
-    "partType": "MainLogicPart"
-}
