+++ GTMBplugin_B/gtmbPlugin/uiScript/function_block_screen.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import mod.client.extraClientApi as clientApi
-from Script_NeteaseMod_TtCLQKGx.function_nodes import nodes
+from gtmbPlugin.function_nodes import nodes
 ViewBinder = clientApi.GetViewBinderCls()
 ViewRequest = clientApi.GetViewViewRequestCls()
 ScreenNode = clientApi.GetScreenNodeCls()
