+++ GTMBplugin_R/ui/enchant.json
@@ -19,7 +19,7 @@
             36
         ]
     },
-    "main": {
+    "main_closable": {
         "type": "screen",
         "controls": [
             {
