+++ GTMBplugin_R/ui/nbteditor.json
@@ -1,5 +1,5 @@
 {
-    "main": {
+    "main_closable": {
         "type": "screen",
         "controls": [
             {
