+++ GTMBplugin_R/ui/structureimport.json
@@ -1,5 +1,5 @@
 {
-    "main": {
+    "main_closable": {
         "type": "screen",
         "controls": [
             {
